pyproject.toml
[tool.poetry]
name = "recipe-blog"
version = "0.1.0"
description = "Blog de receitas, um dashboard para criação de receitas e disponibilizalas para quem quiser"
authors = ["Juan Carlos Leone Barreto <jc.leone98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
markdown = "^3.7"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue . && isort . "
lint-check = "blue --check --diff . && isort --check --diff . "
dev = "python manage.py runserver"